# WIP Optimize rendering by caching renderings when possible.