@zach-klippenstein 2019-12-07
This makes the process of bubbling outputs up the tree slightly slower (have to call each parent to invalidate its rendering cache), so don't merge this until `tick` is benchmarked as well.